Revert "modified the  model health check implementation to capture review feedback"
This reverts commit 98fd6a24806a2977e78928fa68067fbe9bafe099.
