+++ start_server.sh
@@ -49,7 +49,7 @@ if [ ! -d "$ENV_PATH" ]
 then
     apt install -y python3.10-venv
     echo "setting up venv"
-    git clone https://github.com/nakinnubis/pyworker "$SERVER_DIR"
+    git clone https://github.com/vast-ai/pyworker "$SERVER_DIR"
 
     python3 -m venv "$WORKSPACE_DIR/worker-env"
     source "$WORKSPACE_DIR/worker-env/bin/activate"
