+++ workers/tgi/server.py
@@ -1,12 +1,12 @@
 import os
 import logging
-from typing import Union, Type, Dict
+from typing import Union, Type
 import dataclasses
 
-from aiohttp import web, ClientResponse, ClientSession # type: ignore
+from aiohttp import web, ClientResponse # type: ignore
 
 from lib.backend import Backend, LogAction
-from lib.data_types import EndpointHandler, MODELLOADEDSTATUS
+from lib.data_types import EndpointHandler
 from lib.server import start_server
 from .data_types import InputData
 
@@ -40,44 +40,6 @@ class GenerateHandler(EndpointHandler[InputData]):
     def make_benchmark_payload(self) -> InputData:
         return InputData.for_test()
 
-    def health_check(self) -> MODELLOADEDSTATUS:
-        """Simple health check that returns the status enum directly"""
-        # This is a synchronous version that can be used for basic health checks
-        # For now, we'll return READY - in the future this could be enhanced
-        # to do a simple check without requiring async/await
-        return MODELLOADEDSTATUS.READY
-
-    async def model_health_check(self, model_server_url: str) -> Dict[str, str]:
-        """TGI-specific health check implementation"""
-        url = f'{model_server_url}/health'                
-        try:
-            async with ClientSession() as session:
-                async with session.get(url) as health_response:
-                    status_code = health_response.status
-                    if status_code == 200:
-                        message = await health_response.text()
-                        return {'status': MODELLOADEDSTATUS.READY.value, 'reason': message}
-                    elif status_code == 503:
-                        try:
-                            error_response = await health_response.json()
-                            error = error_response.get("error", "")
-                            error_type = error_response.get("error_type", "")
-                            reason = f'{error} {error_type}'.strip()
-                        except Exception:
-                            reason = "Unhealthy (invalid JSON error response)"
-                        return {'status': MODELLOADEDSTATUS.FAILED.value, 'reason': reason}
-                    else:
-                        return {
-                            'status': MODELLOADEDSTATUS.DEFERRED_TO_LOG_FILE.value,
-                            'reason': f'Model health endpoint not ready (status: {status_code})'
-                        }
-        except Exception as e:
-            log.debug(f"Health check exception: {str(e)}")
-            return {
-                'status': MODELLOADEDSTATUS.FAILED.value,
-                'reason': f'Exception during health check: {str(e)}'
-            }
-
     async def generate_client_response(
         self, client_request: web.Request, model_response: ClientResponse
     ) -> Union[web.Response, web.StreamResponse]:
@@ -103,44 +65,6 @@ class GenerateStreamHandler(EndpointHandler[InputData]):
 
     def make_benchmark_payload(self) -> InputData:
         return InputData.for_test()
-
-    def health_check(self) -> MODELLOADEDSTATUS:
-        """Simple health check that returns the status enum directly"""
-        # This is a synchronous version that can be used for basic health checks
-        # For now, we'll return READY - in the future this could be enhanced
-        # to do a simple check without requiring async/await
-        return MODELLOADEDSTATUS.READY
-
-    async def model_health_check(self, model_server_url: str) -> Dict[str, str]:
-        """TGI-specific health check implementation (same as GenerateHandler)"""
-        url = f'{model_server_url}/health'                
-        try:
-            async with ClientSession() as session:
-                async with session.get(url) as health_response:
-                    status_code = health_response.status
-                    if status_code == 200:
-                        message = await health_response.text()
-                        return {'status': MODELLOADEDSTATUS.READY.value, 'reason': message}
-                    elif status_code == 503:
-                        try:
-                            error_response = await health_response.json()
-                            error = error_response.get("error", "")
-                            error_type = error_response.get("error_type", "")
-                            reason = f'{error} {error_type}'.strip()
-                        except Exception:
-                            reason = "Unhealthy (invalid JSON error response)"
-                        return {'status': MODELLOADEDSTATUS.FAILED.value, 'reason': reason}
-                    else:
-                        return {
-                            'status': MODELLOADEDSTATUS.DEFERRED_TO_LOG_FILE.value,
-                            'reason': f'Model health endpoint not ready (status: {status_code})'
-                        }
-        except Exception as e:
-            log.debug(f"Health check exception: {str(e)}")
-            return {
-                'status': MODELLOADEDSTATUS.FAILED.value,
-                'reason': f'Exception during health check: {str(e)}'
-            }
 
     async def generate_client_response(
         self, client_request: web.Request, model_response: ClientResponse
